requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.4
django-import-export==4.3.3
django-unfold==0.43.0
djangorestframework==3.15.2
sqlparse==0.5.3
tablib==3.7.0
tzdata==2024.2
